+++ packages/cli-tools/bin/streamr-storage-node-list-streams.ts
@@ -5,7 +5,7 @@ import StreamrClient from 'streamr-client'
 import { createClientCommand } from '../src/command'
 
 createClientCommand((async (client: StreamrClient, storageNodeAddress: string) => {
-    const { streams } = await client.getStoredStreamsOf(storageNodeAddress)
+    const { streams } = await client.getStoredStreams(storageNodeAddress)
     if (streams.length > 0) {
         console.info(EasyTable.print(streams.map((stream) => {
             return {