+++ packages/client/src/ConfigTest.ts
@@ -4,6 +4,7 @@ function toNumber(value: any): number | undefined {
 
 const sideChainConfig = {
     name: 'streamr',
+    chainId: 8995,
     rpcs: [{
         url: process.env.SIDECHAIN_URL || `http://${process.env.STREAMR_DOCKER_DEV_HOST || '10.200.10.1'}:8546`,
         timeout: toNumber(process.env.TEST_TIMEOUT) ?? 30 * 1000,
@@ -50,6 +51,7 @@ export const ConfigTest = {
     },
     mainChainRPCs: {
         name: 'dev_ethereum',
+        chainId: 8995,
         rpcs: [{
             url: process.env.ETHEREUM_SERVER_URL || `http://${process.env.STREAMR_DOCKER_DEV_HOST || '10.200.10.1'}:8545`,
             timeout: toNumber(process.env.TEST_TIMEOUT) ?? 30 * 1000