+++ packages/client/src/Stream.ts
@@ -235,7 +235,7 @@ class StreamrStream implements StreamMetadata {
     }
 
     async getStorageNodes() {
-        return this._nodeRegistry.getStorageNodesOf(this.id)
+        return this._nodeRegistry.getStorageNodes(this.id)
     }
 
     /**