+++ packages/client/src/publish/PublishPipeline.ts
@@ -186,7 +186,11 @@ export default class PublishPipeline implements Context, Stoppable {
      * Creates a Defer to be resolved when message gets sent to node.
      */
     async publish<T>(publishMetadata: PublishMetadataStrict<T>): Promise<StreamMessage<T>> {
-        this.debug('publish >> %o', publishMetadata)
+        this.debug('publish >> %o', {
+            streamDefinition: formStreamDefinitionDescription(publishMetadata.streamDefinition),
+            timestamp: publishMetadata.timestamp,
+            partitionKey: publishMetadata.partitionKey
+        })
         this.startQueue()
 
         const defer = Defer<StreamMessage<T>>()