+++ packages/client/CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Method `getStorageNodesOf()` renamed to `getStorageNodes()`
+- Method `getStoredStreamsOf()` renamed to `getStoredStreams()`
+- Method `isStreamStoredInStorageNode()` renamed to `isStoredStream()`
 - Method `stream.update()` now requires a parameter `props`
 - Method `unRegisterStorageEventListeners()` renamed to `unregisterStorageEventListeners()`
 
@@ -20,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Remove method `getAllStorageNodes()`, use `getStorageNodes()` without arguments instead
 - Remove (non-functional) client configuration options `autoConnect` and `autoDisconnect`
 - Remove method `disconnect()`, use `destroy()` instead
 - Remove method `unsubscribeAll()`, use `unsubscribe()` without arguments instead
@@ -28,10 +32,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [6.0.2] - 2022-03-04
+
+### Fixed
+
+- Fixed an issue in which method `searchStreams` would throw on invalid metadata (NET-730)
+
+
 ## [6.0.1] - 2022-02-24
 
 ### Fixed
 - Fixed an import so that the client successfully loads in a web browser environment (NET-721)
 
-[Unreleased]: https://github.com/streamr-dev/network-monorepo/compare/client/v6.0.1...HEAD
+[Unreleased]: https://github.com/streamr-dev/network-monorepo/compare/client/v6.0.2...HEAD
+[6.0.2]: https://github.com/streamr-dev/network-monorepo/compare/client/v6.0.1...client/v6.0.2
 [6.0.1]: https://github.com/streamr-dev/network-monorepo/compare/client/v6.0.0...client/v6.0.1